==> Eloise_Clayton/1. Combining Datasets.ipynb <==
import pandas as pd

df1 = pd.read_csv('cross_solvent_solubility_dataset_BSc_2024.csv')
df2 = pd.read_csv('basf_inchi_logS.csv')

df2['solute_inchikey'] = df2['solute_inchikey'].str.replace('InChIKey=', '', regex=False)

df2.rename(columns={'logS_M':'LogS'},inplace=True)

combined_df = pd.concat([df1, df2]).drop_duplicates()

combined_df.to_csv('combined_data.csv', index=False)
