==> Eloise_Clayton/2. Identifying and Averaging Duplicates.ipynb <==
import pandas as pd

df = pd.read_csv('combined_data.csv')
df = df.drop(columns=['solute_smiles', 'S_M', 'ID', "MP_C",'MW'])

#indentify duplicate rows for "logS"
columns_to_check = [col for col in df.columns if col not in ['LogS']]
duplicates = df[df.duplicated(subset=columns_to_check, keep=False)]

duplicates.to_csv('duplicates.csv', index=False)

df_no_duplicates = df.drop(duplicates.index)
df_no_duplicates.to_csv('cd_no_duplicates.csv', index=False)

# Ensure 'LogS' is numeric before averaging
duplicates = duplicates.copy()
duplicates['LogS'] = pd.to_numeric(duplicates['LogS'], errors='coerce')

# Compute the average 'LogS' for duplicate groups
averaged_duplicates = duplicates.groupby(columns_to_check, as_index=False)['LogS'].mean()
averaged_duplicates.to_csv("average_duplicates.csv",index=False)

# Append the averaged rows
df_no_duplicates = pd.concat([df_no_duplicates, averaged_duplicates], ignore_index=True)

df_no_duplicates.to_csv('combined_data_no_duplicates_and_ave_duplicates.csv', index=False)
