==> Eloise_Clayton/4. Duplicate Analysis.ipynb <==
import pandas as pd

df = pd.read_csv("combined_data_no_duplicates_and_ave_duplicates.csv")

# Consider only 'solute_name' and 'solvent_name' columns for duplication check
df_subset = df[['solute_name', 'solvent_name']]

# Count occurrences of each unique combination of solute_name and solvent_name
duplicate_counts = df_subset.value_counts().reset_index()

# Rename the count column
duplicate_counts.columns = ['solute_name', 'solvent_name', 'count']

# Filter and display only rows with counts greater than 1
duplicates_filtered = duplicate_counts[duplicate_counts['count'] > 1]

duplicates_filtered.to_csv("duplicates_filtered.csv", index=False)

import pandas as pd

df = pd.read_csv("combined_data_no_duplicates_and_ave_duplicates.csv")

# Consider only 'solute_name' and 'solvent_name' columns for duplication check
df_subset = df[['solute_inchikey', 'solvent_smiles']]

# Count occurrences of each unique combination of solute_name and solvent_name
duplicate_counts = df_subset.value_counts().reset_index()

# Rename the count column
duplicate_counts.columns = ['solute_inchikey', 'solvent_smiles', 'count']

# Filter and display only rows with counts greater than 1
duplicates_filtered = duplicate_counts[duplicate_counts['count'] > 1]

duplicates_filtered.to_csv("duplicates_2.csv", index=False)

import pandas as pd

def find_duplicates(df):
    duplicate_columns = ["solute_name", "solvent_name", "solvent_smiles"]
    duplicate_counts = df.groupby(duplicate_columns).size().reset_index(name='count')
    duplicate_counts = duplicate_counts[duplicate_counts['count'] > 1]
    return duplicate_counts  


df = pd.read_csv("combined_data_no_duplicates_and_ave_duplicates.csv")
duplicate_results = find_duplicates(df)
duplicate_results.to_csv("duplicates_3.csv", index=False)

import pandas as pd

def find_duplicates(df):
    duplicate_columns = ["solute_name", "solvent_name", "solute_inchikey"]
    duplicate_counts = df.groupby(duplicate_columns).size().reset_index(name='count')
    duplicate_counts = duplicate_counts[duplicate_counts['count'] > 1]
    return duplicate_counts 

df = pd.read_csv("combined_data_no_duplicates_and_ave_duplicates.csv")
duplicate_results = find_duplicates(df)
duplicate_results.to_csv("duplicates_4.csv", index=False)
