==> Eloise_Clayton/5. Identifying SMILES Issues.ipynb <==
import pandas as pd

df = pd.read_csv("combined_data_no_duplicates_and_ave_duplicates.csv")
duplicates = pd.read_csv("duplicates_4.csv")

# Merge to get solvent_name and solute_smiles for the identified duplicates
duplicate_smiles = df.merge(duplicates, on=["solute_name", "solvent_name", "solute_inchikey"], how="inner")

# Select only solvent_name and solute_smiles columns
duplicate_smiles = duplicate_smiles[["solvent_name", "solvent_smiles"]].drop_duplicates()

duplicate_smiles.to_csv("unique_duplicate_smiles.csv", index=False)
