==> Eloise_Clayton/9. Adding Solute Smiles.ipynb <==
import pandas as pd

filtered_df = pd.read_csv("filtered_solvent_data.csv")
combined_df = pd.read_csv("combined_data.csv")

# Remove duplicates in combined_df based on 'solute_inchikey'
combined_df = combined_df.drop_duplicates(subset=['solute_inchikey'])

# Merge filtered_df and combined_df based on 'solute_inchikey'
merged_df = filtered_df.merge(combined_df[['solute_inchikey', 'solute_smiles']], 
                               on='solute_inchikey', 
                               how='left')

merged_df.to_csv("filtered_data_with_smiles.csv", index=False)
