==> Eloise_Clayton/11. Extracting Unique Solvents for EPS.ipynb <==
import pandas as pd

file_path = "filtered_solvent_data.csv"
df = pd.read_csv(file_path)

# Extract unique solvents
unique_solvents = df["solvent_name"].unique()

# Save the unique solvents as a new CSV file
output_file = "unique_solvents.csv"
pd.DataFrame(unique_solvents, columns=["solvent_name"]).to_csv(output_file, index=False)
