==> Eloise_Clayton/16. Unzip.ipynb <==
import zipfile

zip_path = "/home/1a310427-f101-441a-87b0-b753ef497740/output 2/eloise_clayton_gas_output.zip"
extract_to = "/home/1a310427-f101-441a-87b0-b753ef497740/output 2/output_gas"

try:
    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
        zip_ref.extractall(extract_to)
    print("Extraction successful!")
except zipfile.BadZipFile:
    print("Error: The file is not a valid ZIP archive or is corrupted.")
except Exception as e:
    print(f"Unexpected error: {e}")

import zipfile

zip_path = "/home/1a310427-f101-441a-87b0-b753ef497740/output 2/eloise_clayton_output_4 1.zip"
extract_to = "/home/1a310427-f101-441a-87b0-b753ef497740/output 2/output_sol"

try:
    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
        zip_ref.extractall(extract_to)
    print("Extraction successful!")
except zipfile.BadZipFile:
    print("Error: The file is not a valid ZIP archive or is corrupted.")
except Exception as e:
    print(f"Unexpected error: {e}")

import zipfile

zip_path = "/home/1a310427-f101-441a-87b0-b753ef497740/output 2/eloise_clayton_solvents_out.zip"
extract_to = "/home/1a310427-f101-441a-87b0-b753ef497740/output 2/output_solvents"

try:
    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
        zip_ref.extractall(extract_to)
    print("Extraction successful!")
except zipfile.BadZipFile:
    print("Error: The file is not a valid ZIP archive or is corrupted.")
except Exception as e:
    print(f"Unexpected error: {e}")
