==> Eloise_Clayton/18. Determining G.ipynb <==
import pandas as pd

gas_df = pd.read_csv('gas_phase_data.csv')
solution_df = pd.read_csv('solution_phase_data.csv')

# Merge the two DataFrames based on 'InChiKey'
merged_df = pd.merge(gas_df, solution_df, on=['solute_inchikey', 'solvent_name'], suffixes=('_gas', '_sol'))

# Calculate G(gas) and G(sol)
merged_df['G_gas'] = merged_df['H_gas'] - 298 * merged_df['S_gas']
merged_df['G_sol'] = merged_df['H_sol'] - 298 * merged_df['S_sol']

# Calculate G(solvation)
merged_df['DeltaG_sol'] = merged_df['G_sol'] - merged_df['G_gas']

# Create the new DataFrame with the required columns
final_df = merged_df[['solute_inchikey', 'solvent_name', 'G_gas', 'G_sol', 'DeltaG_sol']]

final_df.to_csv('free_energy_data.csv', index=False)
